pyproject.toml
[project]
name = "snackbox"
version = "0.1.0"
description = "A coding efficiency toolbox"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "gitpython>=3.1.44",
    "pytest>=8.3.5",
    "pytest-testmon>=2.1.3",
    "pytest-watch>=4.2.0",
    "typer>=0.15.2",
    ]

[project.scripts]
snack = "snackbox.main:app"


[dependency-groups]
dev = [
    "ruff>=0.9.10",
    ]
